web/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
redis==5.0.1
psycopg2-binary==2.9.9
SQLAlchemy==1.4.22
